refactor : 주문테이블에서 손님수변경, 단체지정에서 빈테이블로 변경시에 상태 체크할 Orders를 CompletionOrders로 변경 상태 validation도 domain 영역으로 이동
